# Optimization - Kerl.java absorb now 600% faster