--- a/src/main/java/com/iota/iri/hash/Kerl.java
+++ b/src/main/java/com/iota/iri/hash/Kerl.java
@@ -13,19 +13,14 @@ public final class Kerl implements Sponge {
     public static final int BIT_HASH_LENGTH = 384;
     public static final int BYTE_HASH_LENGTH = BIT_HASH_LENGTH / 8;
 
-    // radix constant conversion for fast math
     public static final BigInteger RADIX = BigInteger.valueOf(Converter.RADIX);
     private static final int MAX_POWERS_LONG = 40;
     private static final BigInteger[] RADIX_POWERS = IntStream.range(0, MAX_POWERS_LONG + 1).mapToObj(RADIX::pow).toArray(BigInteger[]::new);
 
-    private final byte[] byte_state;
-    private final int[] trit_state;
     private final Keccak.Digest384 keccak;
 
     protected Kerl() {
         this.keccak = new Keccak.Digest384();
-        this.byte_state = new byte[BYTE_HASH_LENGTH];
-        this.trit_state = new int[Sponge.HASH_LENGTH];
     }
 
     @Override
@@ -39,15 +34,11 @@ public final class Kerl implements Sponge {
             throw new RuntimeException("Illegal length: " + length);
         }
         for (int pos = offset; pos < offset + length; pos += HASH_LENGTH) {
-            //copy trits[offset:offset+length]
-            System.arraycopy(trits, pos, trit_state, 0, HASH_LENGTH);
-
-            //convert to bits
-            trit_state[HASH_LENGTH - 1] = 0;
-            bytesFromBigInt(bigIntFromTrits(trit_state, 0, HASH_LENGTH), byte_state);
-
-            //run keccak
-            keccak.update(byte_state);
+            //convert to bytes && update
+            byte[] state = new byte[BYTE_HASH_LENGTH];
+            trits[pos + HASH_LENGTH - 1] = 0;
+            bytesFromBigInt(bigIntFromTrits(trits, pos, HASH_LENGTH), state);
+            keccak.update(state);
         }
     }
 
@@ -58,19 +49,20 @@ public final class Kerl implements Sponge {
         }
         try {
             for (int pos = offset; pos < offset + length; pos += HASH_LENGTH) {
-                this.keccak.digest(byte_state, 0, BYTE_HASH_LENGTH);
-                //convert to trits
-                tritsFromBigInt(bigIntFromBytes(byte_state, 0, BYTE_HASH_LENGTH), trit_state, 0, Sponge.HASH_LENGTH);
 
-                //copy with offset
-                trit_state[HASH_LENGTH - 1] = 0;
-                System.arraycopy(trit_state, 0, trits, pos, HASH_LENGTH);
+                byte[] state = new byte[BYTE_HASH_LENGTH];
+                keccak.digest(state, 0, BYTE_HASH_LENGTH);
+
+                //convert into trits
+                BigInteger value = new BigInteger(state);
+                tritsFromBigInt(value, trits, pos, Sponge.HASH_LENGTH);
+                trits[pos + HASH_LENGTH - 1] = 0;
 
                 //calculate hash again
-                for (int i = byte_state.length; i-- > 0; ) {
-                    byte_state[i] = (byte) (byte_state[i] ^ 0xFF);
+                for (int i = state.length; i-- > 0; ) {
+                    state[i] = (byte) (state[i] ^ 0xFF);
                 }
-                keccak.update(byte_state);
+                keccak.update(state);
             }
         } catch (DigestException e) {
             e.printStackTrace(System.err);
@@ -86,20 +78,15 @@ public final class Kerl implements Sponge {
         }
         BigInteger value = BigInteger.ZERO;
         for (int n = offset + size - 1; n >= offset; ) {
+            int count = 0;
             long num = 0L;
-            int power = 0;
-            while (n >= offset && power < MAX_POWERS_LONG) {
+            while (n >= offset && count < MAX_POWERS_LONG) {
                 num = 3 * num + trits[n--];
-                power++;
+                count++;
             }
-            value = value.multiply(RADIX_POWERS[power]).add(BigInteger.valueOf(num));
+            value = value.multiply(RADIX_POWERS[count]).add(BigInteger.valueOf(num));
         }
         return value;
-    }
-
-    public static BigInteger bigIntFromBytes(final byte[] bytes, final int offset, final int size) {
-        return (bytes.length == offset + size) ? new BigInteger(bytes)
-            : new BigInteger(Arrays.copyOfRange(bytes, offset, offset + size));
     }
 
     public static void tritsFromBigInt(final BigInteger value, final int[] destination, final int offset, final int size) {
@@ -130,10 +117,11 @@ public final class Kerl implements Sponge {
         if (destination.length < BYTE_HASH_LENGTH) {
             throw new IllegalArgumentException("Destination array has invalid size.");
         }
-        int start = BYTE_HASH_LENGTH - (value.bitLength() / 8 + 1);
+        byte[] bytes = value.toByteArray();
+        int start = BYTE_HASH_LENGTH - bytes.length;
         Arrays.fill(destination, 0, start, (byte) (value.signum() < 0 ? -1 : 0));
-        for (byte aByte : value.toByteArray()) {
-            destination[start++] = aByte;
+        for (int i = 0; i < bytes.length; i++) {
+            destination[start++] = bytes[i];
         }
     }
 }