Hash routine in Kerl.java absorb(...) now 600% faster!
Snapshot initial now loading 368% faster.

Big time hog was bigIntFromTrits(...)
- Created a fast track for Radix power multiplication.
- Created a native math routine to munch 36 trits at a time into the main math routine.
- Added test to ensure that there is now math overflow.

LOADING SNAPSHOT TIMES
INITIAL      BEFORE    AFTER 
ABSORB        7920    1306 ms
SNAPSHOT   9693    2879 ms
AVG 20 rounds
ABSORB        7195    1197 ms
SNAPSHOT   8760    2378 ms

Signed-off-by: footloosejava <32090281+footloosejava@users.noreply.github.com>
Fixes # (issue)

## Type of change
- Enhancement - Speeds up Kerl considerably

# How Has This Been Tested?
- Test written
- Tested on Snapshot loading
- Tested via existing Kerl and Curl and Bundle Validator tests

# Checklist:
 [X] My code follows the style guidelines for this project
 [X] I have performed a self-review of my own code
 [X] I have added tests that prove my fix is effective or that my feature works
 [X] New and existing unit tests pass locally with my changes