@th0br0 2018-04-13
Did you happen to also compare the old implementation (which is known to be slow) with the optimised variant that all libraries are using?
See here for a Java, BigInt-less implementation:
https://github.com/iotaledger/iota.lib.java/blob/master/jota/src/main/java/jota/pow/Kerl.java