@footloosejava 2018-04-13
@th0br0 Didn't know about the others.
I just took a look. **Mine is a lot simpler and more straightforward.** I'll have to see how they compare ....