@th0br0 2018-04-13
It definitely is simpler and easier to parse. 
The reason for the current IRI codebase (still) doing things slowly mainly was prudence.