@footloosejava 2018-04-13
@th0br0 test done. Mine 'absorb' is is **at least 430% faster** than the other one. Over 5 seconds faster each round.
JCURL 'squeeze' is is **178% faster** than mine. However, the difference from 'squeeze' on digesting a 20 MB file is less than 1/3 of 1/1000th of a second.

Tested digesting 20 complete load of of the 20MB+ snapshot.

FIRST RUN | OLD | MINE | JCURL
-----------  | ----- | ------ | -------
SNAPSHOT |  9693	| **2879** | 7487 millis
ABSORB | 7920 |  **1306** | 5572 millis
SQUEEZE|	    800828|	832151|	367134 nanos (1/3 of a millis)

AVG of 20 |   OLD |   MINE |  JCURL
-------- | -------- | -----| -------
SNAPSHOT|	8760 |	**2378**	|6754
ABSORB|		7195 |	**1197**|	5158
SQUEEZE	 |   187030|	186910	|332769 nanos (1/3 of a millis)