@footloosejava 2018-04-14
Just realized that simpler is even better. Removed all the complicated goop, and kept the essential goodness .. and pretty much as fast as before but 1000% easier to read.