@footloosejava 2018-04-14
@pinpong @jpkrohling In case you are interested.