@gjeee 2018-04-14
nice job again footloose! i used your Kerl class and my CPU sure seems to like it......