@footloosejava 2018-04-14
Commit incoming - I just found a way to make it faster again ... and **even simpler** with **no instance state**. There will be lots of ice-cream with cappuccinos tonight ☕️  + 🍨  :)