@alon-e 2018-04-15
Hey @footloosejava - thanks for this contribution!
reviewing now.
did you also run the `test_vectors` in https://github.com/iotaledger/kerl ?