@gjeee 2018-04-15
@footloosejava i applied your latest Kerl.java commit as well but i think this one has issues....my CPU spikes as never before.....the one before was like baby smooth......

????