@footloosejava 2018-04-16
@GalRogozinski see last commit and comments. Thanks!