@footloosejava 2018-04-16
@gjeee No problem. Your feedback and second set of eyes is always appreciated!