@alon-e 2018-05-02
@footloosejava @galrogo I ran the test_vectors locally from: https://github.com/iotaledger/kerl 
all pass as expected :+1: